refactor: 점포에 대한 점주의 권한 검증 #73

+++ src/main/java/org/example/luckyburger/domain/shop/dto/response/ShopDashboardResponse.java
@@ -5,13 +5,13 @@ import lombok.Builder;
 @Builder
 public record ShopDashboardResponse(
         Long shopId,
-        Integer todayOrderCount,
+        Long todayOrderCount,
         Long todayTotalSales,
         Double averageRating
 ) {
     public static ShopDashboardResponse of(
             Long shopId,
-            Integer todayOrderCount,
+            Long todayOrderCount,
             Long todayTotalSales,
             Double averageRating) {
         return ShopDashboardResponse.builder()
